refactor: split out TS files into JS + declaration files
