+++ lib/validators/schemas/json-schema.js
@@ -1,0 +1,4 @@
+// TODO: delete
+
+Object.defineProperty(exports, '__esModule', { value: true });
+// # sourceMappingURL=json-schema.js.map
